+++ receipt-analyzer-service/app/src/main/java/pl/madzierski/daniel/receipt/model/UpdateRevisionResponse.java
@@ -2,6 +2,7 @@ package pl.madzierski.daniel.receipt.model;
 
 import pl.madzierski.daniel.receipt.ReceiptResolverStrategyType;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.Collection;
 import java.util.Set;
@@ -12,7 +13,7 @@ public record UpdateRevisionResponse(
     ReceiptResolverStrategyType resolver,
     LocalDateTime createdDate,
     String brand,
-    Double totalPrice,
+    BigDecimal totalPrice,
     LocalDateTime payingDate,
     String address,
     Boolean isPreferredRevision,
@@ -54,10 +55,10 @@ public record UpdateRevisionResponse(
     public record ItemResponse(
         String id,
         String name,
-        Double amount,
-        Double unitPrice,
-        Double discount,
-        Double totalPrice,
+        BigDecimal amount,
+        BigDecimal unitPrice,
+        BigDecimal discount,
+        BigDecimal totalPrice,
         Integer position
     ) {
     }
